+++ src/polymer/elements/rules.ts
@@ -14,4 +14,5 @@
 
 import './iron-flex-layout-classes';
 import './iron-flex-layout-import';
+import './iron-form-v1-to-v2';
 import './paper-toolbar-v1-to-v2';