+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+- Adds the new rule `iron-form-v1-to-v2`, which warns about `iron-form` being used as a type extension of `<form>`.
+  - The fix automatically wraps the `<form is="iron-form">` into an `<iron-form>` and moves `id`, `with-credentials`, `headers` properties and `on-iron-form-*` events from `<form>` to `<iron-form>`.
 
 ## [3.0.0-pre.3] - 2018-03-07