+++ CHANGELOG.md
@@ -5,14 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+<!-- ## Unreleased -->
+<!-- Add new, unreleased changes here. -->
+
+## [3.0.0-pre.14] - 2018-03-09
 * [breaking] JavascriptDocument#ast is now a `babel.File` rather than a
   `babel.Program`. Use `jsDoc.ast.program` instead of `jsDoc.ast` in the
   unlikely case that the `Program` is required.
 * Fix bug where if a package's name was a prefix of one of its dependencies,
   that dependency would not resolve to its components directory.
-
-<!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.13] - 2018-03-05
  * Support specifying tag names in jsdoc, e.g. `@customElement fancy-button`